[vm] Reorganize rules into categories
