+++ pmd-vm/src/main/java/net/sourceforge/pmd/lang/vm/rule/design/ExcessiveTemplateLengthRule.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.vm.rule.basic;
+package net.sourceforge.pmd.lang.vm.rule.design;
 
 import net.sourceforge.pmd.lang.vm.ast.ASTprocess;
 import net.sourceforge.pmd.lang.vm.rule.AbstractStatisticalVmRule;
