+++ pmd-vm/src/main/java/net/sourceforge/pmd/lang/vm/rule/errorprone/EmptyIfStmtRule.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.vm.rule.basic;
+package net.sourceforge.pmd.lang.vm.rule.errorprone;
 
 import org.apache.commons.lang3.StringUtils;
