+++ pmd-vm/src/main/resources/category/vm/bestpractices.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0"?>
+
+<ruleset name="Best Practices"
+         xmlns="http://pmd.sourceforge.net/ruleset/2.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://pmd.sourceforge.net/ruleset/2.0.0 http://pmd.sourceforge.net/ruleset_2_0_0.xsd">
+
+    <description>
+Rules which enforce generally accepted best practices.
+    </description>
+
+    <rule name="AvoidReassigningParameters"
+          since="5.1"
+          message="Avoid reassigning macro parameters such as ''{0}''"
+          class="net.sourceforge.pmd.lang.vm.rule.bestpractices.AvoidReassigningParametersRule"
+          externalInfoUrl="${pmd.website.baseurl}/pmd_rules_vm_bestpractices.html#avoidreassigningparameters">
+        <description>
+Reassigning values to incoming parameters is not recommended.  Use temporary local variables instead.
+        </description>
+        <priority>2</priority>
+    </rule>
+
+    <rule name="UnusedMacroParameter"
+          since="5.1"
+          message="Avoid unused macro parameters such as ''{0}''"
+          class="net.sourceforge.pmd.lang.vm.rule.bestpractices.UnusedMacroParameterRule"
+          externalInfoUrl="${pmd.website.baseurl}/pmd_rules_vm_bestpractices.html#unusedmacroparameter">
+        <description>
+Avoid unused macro parameters. They should be deleted.
+        </description>
+        <priority>2</priority>
+    </rule>
+
+</ruleset>
