+++ pmd-vm/src/main/resources/rulesets/vm/basic.xml
@@ -9,116 +9,16 @@
 The Basic Velocity ruleset contains basic rules for Apache Velocity pages.
   </description>
 
-  <rule name="AvoidDeeplyNestedIfStmts"
-        since="5.1"
-        message="Deeply nested if..then statements are hard to read"
-        class="net.sourceforge.pmd.lang.vm.rule.basic.AvoidDeeplyNestedIfStmtsRule"
-        externalInfoUrl="${pmd.website.baseurl}/pmd_rules_vm_basic.html#avoiddeeplynestedifstmts">
-    <description>
-Avoid creating deeply nested if-then statements since they are harder to read and error-prone to maintain.
-    </description>
-    <priority>3</priority>
-  </rule>
+    <rule ref="category/vm/design.xml/AvoidDeeplyNestedIfStmts" deprecated="true" />
+    <rule ref="category/vm/design.xml/CollapsibleIfStatements" deprecated="true" />
+    <rule ref="category/vm/design.xml/ExcessiveTemplateLength" deprecated="true" />
+    <rule ref="category/vm/design.xml/NoInlineJavaScript" deprecated="true" />
+    <rule ref="category/vm/design.xml/NoInlineStyles" deprecated="true" />
 
-  <rule name="CollapsibleIfStatements"
-        since="5.1"
-        message="These nested if statements could be combined"
-        class="net.sourceforge.pmd.lang.vm.rule.basic.CollapsibleIfStatementsRule"
-        externalInfoUrl="${pmd.website.baseurl}/pmd_rules_vm_basic.html#collapsibleifstatements">
-    <description>
-Sometimes two consecutive 'if' statements can be consolidated by separating their conditions with a boolean short-circuit operator.
-    </description>
-    <priority>3</priority>
-  </rule>
+    <rule ref="category/vm/bestpractices.xml/AvoidReassigningParameters" deprecated="true" />
+    <rule ref="category/vm/bestpractices.xml/UnusedMacroParameter" deprecated="true" />
 
-  <rule name="ExcessiveTemplateLength"
-        since="5.1"
-        message="Template is too long"
-        class="net.sourceforge.pmd.lang.vm.rule.basic.ExcessiveTemplateLengthRule"
-        externalInfoUrl="${pmd.website.baseurl}/pmd_rules_vm_basic.html#excessivetemplatelength">
-    <description>
-The template is too long. It should be broken up into smaller pieces.
-    </description>
-    <priority>3</priority>
-    <properties>
-        <property name="minimum" value="1000"/>
-    </properties>
-  </rule>
-
-  <rule name="AvoidReassigningParameters"
-        since="5.1"
-        message="Avoid reassigning macro parameters such as ''{0}''"
-        class="net.sourceforge.pmd.lang.vm.rule.basic.AvoidReassigningParametersRule"
-        externalInfoUrl="${pmd.website.baseurl}/pmd_rules_vm_basic.html#avoidreassigningparameters">
-    <description>
-Reassigning values to incoming parameters is not recommended.  Use temporary local variables instead.
-    </description>
-    <priority>2</priority>
-  </rule>
-
-  <rule name="EmptyIfStmt"
-        since="5.1"
-        message="Avoid empty if statements"
-        class="net.sourceforge.pmd.lang.vm.rule.basic.EmptyIfStmtRule"
-        externalInfoUrl="${pmd.website.baseurl}/pmd_rules_vm_basic.html#emptyifstmt">
-    <description>
-Empty if statements should be deleted.
-    </description>
-    <priority>2</priority>
-  </rule>
-
-  <rule name="EmptyForeachStmt"
-        since="5.1"
-        message="Avoid empty foreach loops"
-        class="net.sourceforge.pmd.lang.vm.rule.basic.EmptyForeachStmtRule"
-        externalInfoUrl="${pmd.website.baseurl}/pmd_rules_vm_basic.html#emptyforeachstmt">
-    <description>
-Empty foreach statements should be deleted.
-    </description>
-    <priority>2</priority>
-  </rule>
-
-  <rule name="UnusedMacroParameter"
-        since="5.1"
-        message="Avoid unused macro parameters such as ''{0}''"
-        class="net.sourceforge.pmd.lang.vm.rule.basic.UnusedMacroParameterRule"
-        externalInfoUrl="${pmd.website.baseurl}/pmd_rules_vm_basic.html#unusedmacroparameter">
-    <description>
-Avoid unused macro parameters. They should be deleted.
-    </description>
-    <priority>2</priority>
-  </rule>
-
-  <rule name="NoInlineJavaScript"
-        since="5.1"
-        message="Avoid inline JavaScript"
-        class="net.sourceforge.pmd.lang.vm.rule.basic.NoInlineJavaScriptRule"
-        externalInfoUrl="${pmd.website.baseurl}/pmd_rules_vm_basic.html#noinlinejavascript">
-    <description>
-Avoid inline JavaScript. Import .js files instead.
-    </description>
-    <priority>2</priority>
-  </rule>
-
-  <rule name="NoInlineStyles"
-        since="5.1"
-        message="Avoid inline styles"
-        language="vm" 
-        class="net.sourceforge.pmd.lang.rule.XPathRule"
-        externalInfoUrl="${pmd.website.baseurl}/pmd_rules_vm_basic.html#noinlinestyles">
-    <description>
-Avoid inline styles. Use css classes instead.
-    </description>
-    <priority>2</priority>
-    <properties>
-        <property name="xpath">
-            <value>
-<![CDATA[
-//Text[matches(@literal, "<[^>]+\s[sS][tT][yY][lL][eE]\s*=")]
-]]>
-            </value>
-        </property>
-    </properties>
-  </rule>
+    <rule ref="category/vm/errorprone.xml/EmptyForeachStmt" deprecated="true" />
+    <rule ref="category/vm/errorprone.xml/EmptyIfStmt" deprecated="true" />
 
 </ruleset>
