+++ pmd-vm/src/main/resources/rulesets/vm/rulesets.properties
@@ -2,4 +2,16 @@
 # BSD-style license; for more info see http://pmd.sourceforge.net/license.html
 #
 
-rulesets.filenames=rulesets/vm/basic.xml
+rulesets.filenames=\
+    category/vm/bestpractices.xml,\
+    category/vm/design.xml,\
+    category/vm/errorprone.xml
+
+#
+# categories without rules
+#
+# category/vm/codestyle.xml
+# category/vm/documentation.xml
+# category/vm/multithreading.xml
+# category/vm/performance.xml
+# category/vm/security.xml
