+++ pmd-vm/src/test/resources/net/sourceforge/pmd/lang/vm/rule/design/xml/CollapsibleIfStatements.xml
