+++ pmd-vm/src/test/resources/net/sourceforge/pmd/lang/vm/rule/errorprone/xml/EmptyForeachStmt.xml
